@@ -1,20 +1,21 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-        <PackageVersion>0.1.3.0</PackageVersion>
-        <AssemblyVersion>0.1.3.0</AssemblyVersion>
-        <FileVersion>0.1.3.0</FileVersion>
+        <TargetFramework>net6.0</TargetFramework>
+        <PackageVersion>0.1.4.0</PackageVersion>
+        <AssemblyVersion>0.1.4.0</AssemblyVersion>
+        <FileVersion>0.1.4.0</FileVersion>
         <Nullable>enable</Nullable>
         <Authors>SubZeroPL</Authors>
         <PackageProjectUrl>https://github.com/SubZeroPL/Jellyfin.Plugin.IvInfo</PackageProjectUrl>
+        <LangVersion>10</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="HtmlAgilityPack" Version="1.11.42"/>
-        <PackageReference Include="Jellyfin.Controller" Version="10.7.7"/>
-        <PackageReference Include="Jellyfin.Model" Version="10.7.7"/>
-        <PackageReference Include="Microsoft.Extensions.Http" Version="5.0.0"/>
+        <PackageReference Include="HtmlAgilityPack" Version="1.11.43"/>
+        <PackageReference Include="Jellyfin.Controller" Version="10.8.0"/>
+        <PackageReference Include="Jellyfin.Model" Version="10.8.0"/>
+        <PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0"/>
     </ItemGroup>
 
     <ItemGroup>
