@@ -38,7 +38,7 @@ namespace Jellyfin.Plugin.IvInfo
 
         public override PluginInfo GetPluginInfo()
         {
-            return new PluginInfo(IvInfoConstants.Name, new Version(0, 1, 2, 0), "Idol Video metadata provider", Id,
+            return new PluginInfo(IvInfoConstants.Name, new Version(0, 1, 4, 0), "Idol Video metadata provider", Id,
                 true);
         }
     }
